Python files in secondary school
This is what I did during my time in the secondary school. At that time, I was not very professional at English so some of the names are not properly named.
Now, I am trying to write codes in a more readable way because I have forgot some of the purposes.